+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.0.0
 
+* Add `guardDijkstraFeaturesForPlutusV1toV3`
+* Add `DirectDepositsNotSupported` and `AccountBalanceIntervalsNotSupported` constructors to `DijkstraContextError`
 * Add `SubTxContextError`
 * Add `DijkstraStAnnTx`
 * Add `scriptsProvidedDijkstraStAnnTx`